🚚 utility/lookup => dns/lookup
