+++ src/plugins/commands/dns/modules/index.ts
@@ -1,0 +1,3 @@
+import lookup from "./lookup";
+
+export default { lookup };
