+++ src/plugins/commands/utility/modules/index.ts
@@ -1,11 +1,9 @@
 import avatar from "./avatar";
 import about from "./about";
-import lookup from "./lookup";
 import stats from "./stats";
 
 export default {
   avatar,
   about,
-  lookup,
   stats,
 };
